requirements.txt
# ── UI ────────────────────────────────────────────────────────────────────────
streamlit>=1.35.0
 
# ── Web Scraper ───────────────────────────────────────────────────────────────
crawl4ai>=0.4.0
playwright>=1.44.0
 
# ── AI / LLM ─────────────────────────────────────────────────────────────────
google-generativeai>=0.7.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-google-genai>=1.0.0
 
# ── Data ─────────────────────────────────────────────────────────────────────
pandas>=2.0.0
 
# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0